LayoutTests/discovery/RUIHPL-7.9.3.1.2.b-expected.txt

Tests DLNA Player for AV Media Format. RUIHPL-7.9.3.1.2: MPEG_PS_NTSC MP2@ML AC3

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Video Track is playable: /home/gar/MediaContent/B-MP2PS_N-3.mpg
PASS Video Track is playing.
PASS Video Track is paused.
PASS successfullyParsed is true

TEST COMPLETE
